+++ packages/fxa-settings/src/components/LinkExpired/index.tsx
@@ -67,13 +67,11 @@ const LinkExpired = ({ linkType }: LinkExpiredProps) => {
       <FtlMsg id={templateValues.messageId}>
         <p className="mt-4 text-sm">{templateValues.messageText}</p>
       </FtlMsg>
-      <div className="flex">
-        <FtlMsg id="resend-link">
-          <button onClick={resendLinkHandler} className="link-blue mt-4">
-            Receive new link
-          </button>
-        </FtlMsg>
-      </div>
+      <FtlMsg id="resend-link">
+        <button onClick={resendLinkHandler} className="link-blue mt-4">
+          Receive new link
+        </button>
+      </FtlMsg>
     </>
   );
 };
