+++ packages/fxa-settings/src/pages/ResetPassword/CompleteResetPassword/en.ftl
