+++ packages/fxa-settings/src/pages/ResetPassword/CompleteResetPassword/index.tsx
@@ -5,15 +5,16 @@
 import React, { useCallback, useState } from 'react';
 import { RouteComponentProps, useNavigate } from '@reach/router';
 import { useForm } from 'react-hook-form';
-import { usePageViewEvent } from '../../lib/metrics';
+import { usePageViewEvent } from '../../../lib/metrics';
 import { FtlMsg } from 'fxa-react/lib/utils';
-import { useFtlMsgResolver } from '../../models/hooks';
-import { useAlertBar } from '../../models';
-import WarningMessage from '../../components/WarningMessage';
-import LinkRememberPassword from '../../components/LinkRememberPassword';
-import LinkExpired from '../../components/LinkExpired';
-import LinkDamaged from '../../components/LinkDamaged';
-import FormResetPasswordWithBalloon from '../../components/FormResetPasswordWithBalloon';
+
+import { useFtlMsgResolver } from '../../../models/hooks';
+import { useAlertBar } from '../../../models';
+import WarningMessage from '../../../components/WarningMessage';
+import LinkRememberPassword from '../../../components/LinkRememberPassword';
+import LinkExpired from '../../../components/LinkExpired';
+import LinkDamaged from '../../../components/LinkDamaged';
+import FormResetPasswordWithBalloon from '../../../components/FormResetPasswordWithBalloon';
 
 // The equivalent complete_reset_password mustache file included account_recovery_reset_password
 // For React, we have opted to separate these into two pages to align with the routes.
