+++ packages/fxa-settings/src/pages/ResetPassword/ConfirmResetPassword/mocks.tsx
